Refactor benchmark code and add script to plot model comparison
