+++ tests/update_benchmarks.py
@@ -58,9 +58,13 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description="Update benchmark tables in README.md and benchmark.rst from CSV"
     )
-    parser.add_argument("--csv", default="benchmark.csv", help="Path to benchmark.csv")
+    parser.add_argument(
+        "--csv", default="outputs/results.csv", help="Path to benchmark.csv"
+    )
     parser.add_argument("--md", default="../README.md", help="Path to README.md")
-    parser.add_argument("--rst", default="benchmark.rst", help="Path to benchmark.rst")
+    parser.add_argument(
+        "--rst", default="../docs/benchmark.rst", help="Path to benchmark.rst"
+    )
     args = parser.parse_args()
 
     main(args.csv, args.md, args.rst)
